+++ libs/core/src/lib/shellbar/shellbar.component.spec.ts
@@ -21,4 +21,26 @@ describe('ShellbarComponent', () => {
     it('should create', () => {
         expect(component).toBeTruthy();
     });
+
+    it('should apply class modifiers', async () => {
+        await fixture.whenStable();
+        let productElementClass =
+            fixture.debugElement.nativeElement.querySelector('.fd-shellbar__group--product')?.classList;
+        expect(productElementClass).not.toContain('fd-shellbar__group--shrink');
+        expect(productElementClass).not.toContain('fd-shellbar__group--basis-auto');
+        component.groupFlex = {
+            product: {
+                shrink: true,
+                flexBasisAuto: true
+            }
+        };
+
+        await fixture.whenStable();
+
+        productElementClass =
+            fixture.debugElement.nativeElement.querySelector('.fd-shellbar__group--product')?.classList;
+
+        expect(productElementClass).toContain('fd-shellbar__group--shrink');
+        expect(productElementClass).toContain('fd-shellbar__group--basis-auto');
+    });
 });