+++ libs/core/src/lib/shellbar/shellbar.component.ts
@@ -20,6 +20,7 @@ import { FD_BUTTON_COMPONENT } from '@fundamental-ngx/core/button';
 import { ComboboxInterface, FD_COMBOBOX_COMPONENT } from '@fundamental-ngx/core/combobox';
 import { ContentDensityMode, contentDensityObserverProviders } from '@fundamental-ngx/core/content-density';
 import { SearchComponent } from '@fundamental-ngx/core/shared';
+import equal from 'fast-deep-equal';
 import { BehaviorSubject, debounceTime, distinctUntilChanged, Subscription, takeUntil } from 'rxjs';
 import { ShellbarActionsComponent } from './shellbar-actions/shellbar-actions.component';
 import { FD_SHELLBAR_SEARCH_COMPONENT } from './tokens';
@@ -29,11 +30,26 @@ export type ShellbarSizes = 's' | 'm' | 'l' | 'xl';
 
 export type Breakpoints = Record<ShellbarSizes, number>;
 
+export type ShellbarGroup = 'product' | 'search' | 'actions';
+
 export interface NormalizedBreakpoint {
     size: ShellbarSizes;
     min: number;
     max: number;
 }
+
+export type ShellbarGroupFlexOptions = Partial<{
+    [key in ShellbarGroup]: {
+        /**
+         * Whether to shrink group.
+         */
+        shrink?: boolean;
+        /**
+         * Whether to apply flex-basis: auto style.
+         */
+        flexBasisAuto?: boolean;
+    };
+}>;
 
 /**
  * The shellbar offers consistent, responsive navigation across all products and applications.
@@ -82,6 +98,26 @@ export class ShellbarComponent implements AfterContentInit, AfterViewInit, OnDes
      */
     @Input()
     sideNav: boolean | SideNavigationInterface = false;
+
+    /**
+     * Shellbar group flex configuration.
+     */
+    @Input()
+    set groupFlex(value: Nullable<ShellbarGroupFlexOptions>) {
+        if (equal(value, this._groupFlex)) {
+            return;
+        }
+
+        this._groupFlex = value;
+        this._cd.detectChanges();
+    }
+
+    get groupFlex(): Nullable<ShellbarGroupFlexOptions> {
+        return this._groupFlex;
+    }
+
+    /** @hidden */
+    private _groupFlex: Nullable<ShellbarGroupFlexOptions>;
 
     /** @hidden */
     @ContentChild(FD_COMBOBOX_COMPONENT, { static: false })